create store_with_historical_data_support class directly and remove module
